+++ theme/CHANGES.md
@@ -2,6 +2,10 @@
 
 ## Next
 
+## 1.2.0
+
+- Feature: Created Members post type
+
 ## 1.1.0
 
 - Feature: Stick call-to-action block on mobile